Fix shallowEqualArrays()
