+++ chartlets.js/packages/lib/src/utils/shallowEqualArrays.ts
@@ -1,0 +1,20 @@
+export function shallowEqualArrays(
+  arr1: unknown[] | undefined,
+  arr2: unknown[] | undefined,
+): boolean {
+  if (arr1 === arr2) {
+    return true;
+  }
+  if (!arr1 || !arr2) {
+    return false;
+  }
+  if (arr1.length !== arr2.length) {
+    return false;
+  }
+  for (let i = 0; i < arr1.length; i++) {
+    if (!Object.is(arr1[i], arr2[i])) {
+      return false;
+    }
+  }
+  return true;
+}
